Another major update. This commit contains now real support for the Nose-Hoover thermostat (activated using 'thermostat nose-hoover') and the implementation of the velocity-verlet integration scheme (activated using 'integrator velocity-verlet'), both written by Alex. The major merge was caused by back-merging the changes from the new group contributions into alex branch (again ...). There are also new tests for the thermostats and integrators, so you can compare them if this is needed.

+++ src/topo.f90
@@ -92,7 +92,7 @@ character(len=256)			::	prm_file
 	real(8), allocatable				::	xtop(:)			! topology/coordinates
 	integer(TINY), allocatable	::	iac(:)			! integer atom codes
 	logical, allocatable				::	heavy(:)		! boolean flag, true if atom >= He
-	real, allocatable						::	crg(:)			! charges
+	real(8), allocatable						::	crg(:)			! charges
 	integer(AI), allocatable		::	cgpatom(:)		! charge groups
 	integer, parameter					::	SOLVENT_SPC=0, SOLVENT_3ATOM=1, SOLVENT_GENERAL=2
 	integer											::	solvent_type
@@ -361,7 +361,7 @@ subroutine topo_reallocate(oldatoms, atoms, waters)
 !locals
 	integer						::	oldbonds, bonds
 	integer						::	oldangles, angles
-	real, allocatable			::	r4temp(:)
+	real(8), allocatable			::	r4temp(:)
 	integer(1), allocatable		::	i1temp(:)
 	type(BOND_TYPE), allocatable::	bndtemp(:)
 	type(ANG_TYPE), allocatable	::	angtemp(:)
